fix: process longer comments (#137)
This is the first step in attempting to solve the case when diff
comments are too long. From testing it looks like the comment limit that
the API returns (65535) is not the real limit, but it is hard to find
any concrete information on what the real limit is. I was able to post
some comments with a length > 150000.

Instead of doing any validation on our end and failing early, this PR
updates the logic to rely on the GH API failure message. We will always
attempt to comment and catch any "Body too long" errors that occur.

This should increase the size of the comment allowed and may fix the
issue for some users. I will follow up this with additional PRs fixing
the issue completely.

re https://github.com/corymhall/cdk-diff-action/issues/34